+++ conversions.hh
@@ -21,13 +21,13 @@ class conversion {
 	public:
 	private:
 	std::string units_file_;
-	std::map<std::string,float> length_units_;
-	std::map<std::string,float> time_units_;
-	std::map<std::string,float> density_units_;
-	std::map<std::string,float> diffusion_constant_units_;
-	std::map<std::string,float> area_units_;
-	std::map<std::string,float> volume_units_;
-	std::map<std::string,float> contamination_units_;
+	std::map<std::string,double> length_units_;
+	std::map<std::string,double> time_units_;
+	std::map<std::string,double> density_units_;
+	std::map<std::string,double> diffusion_constant_units_;
+	std::map<std::string,double> area_units_;
+	std::map<std::string,double> volume_units_;
+	std::map<std::string,double> contamination_units_;
 	std::map<std::string,int> dimensions_;
 	protected:
 	public:
@@ -37,20 +37,20 @@ class conversion {
 	//getters
 	inline std::string units_file() { return units_file_; }
 	
-	float convert_length_from(std::string _unit, float _length);
-	float convert_length_to(std::string _unit, float _length);
-	float convert_time_from(std::string _unit, float _time);
-	float convert_time_to(std::string _unit, float _time);
-	float convert_density_from(std::string _unit, float _density);
-	float convert_density_to(std::string _unit, float _density);
-	float convert_diffusion_constant_from(std::string _unit, float _diffusion_constant);
-	float convert_diffusion_constant_to(std::string _unit, float _diffusion_constant);
-	float convert_area_from(std::string _unit, float _area);
-	float convert_area_to(std::string _unit, float _area);
-	float convert_volume_from(std::string _unit, float _volume);
-	float convert_volume_to(std::string _unit, float _volume);
-	float convert_contamination_from(std::string _unit, float _contamination);
-	float convert_contamination_to(std::string _unit, float _contamination);
+	double convert_length_from(std::string _unit, double _length);
+	double convert_length_to(std::string _unit, double _length);
+	double convert_time_from(std::string _unit, double _time);
+	double convert_time_to(std::string _unit, double _time);
+	double convert_density_from(std::string _unit, double _density);
+	double convert_density_to(std::string _unit, double _density);
+	double convert_diffusion_constant_from(std::string _unit, double _diffusion_constant);
+	double convert_diffusion_constant_to(std::string _unit, double _diffusion_constant);
+	double convert_area_from(std::string _unit, double _area);
+	double convert_area_to(std::string _unit, double _area);
+	double convert_volume_from(std::string _unit, double _volume);
+	double convert_volume_to(std::string _unit, double _volume);
+	double convert_contamination_from(std::string _unit, double _contamination);
+	double convert_contamination_to(std::string _unit, double _contamination);
 	void initialise();
 
 };
