+++ vector_ops.cc
@@ -1,13 +1,13 @@
 #include "vector_ops.hh"
 
-void operator+=(std::vector<float>& u, const std::vector<float>& v) { u = u+v; }
-void operator-=(std::vector<float>& u, const std::vector<float>& v) { u = u-v; }
-void operator*=(std::vector<float>& u, const std::vector<float>& v) { u = u*v; }
-void operator*=(std::vector<float>& u, const float& v) { u = u*v; }
-void operator*=(const float& v, std::vector<float>& u) { u = u*v; }
+void operator+=(std::vector<double>& u, const std::vector<double>& v) { u = u+v; }
+void operator-=(std::vector<double>& u, const std::vector<double>& v) { u = u-v; }
+void operator*=(std::vector<double>& u, const std::vector<double>& v) { u = u*v; }
+void operator*=(std::vector<double>& u, const double& v) { u = u*v; }
+void operator*=(const double& v, std::vector<double>& u) { u = u*v; }
 
-std::vector<float> operator+(const std::vector<float>& u, const std::vector<float>& v) {
-	std::vector<float> w;
+std::vector<double> operator+(const std::vector<double>& u, const std::vector<double>& v) {
+	std::vector<double> w;
 	w.resize(u.size());
 	for (int i = 0; i < u.size(); i++) {
 		w[i] = u[i];
@@ -15,8 +15,8 @@ std::vector<float> operator+(const std::vector<float>& u, const std::vector<floa
 	}
 	return w;
 }
-std::vector<float> operator-(const std::vector<float>& u, const std::vector<float>& v) {
-	std::vector<float> w;
+std::vector<double> operator-(const std::vector<double>& u, const std::vector<double>& v) {
+	std::vector<double> w;
 	w.resize(u.size());
 	for (int i = 0; i < u.size(); i++) {
 		w[i] = u[i];
@@ -24,8 +24,8 @@ std::vector<float> operator-(const std::vector<float>& u, const std::vector<floa
 	}
 	return w;
 }
-std::vector<float> operator*(const std::vector<float>& u, const std::vector<float>& v) {
-	std::vector<float> w;
+std::vector<double> operator*(const std::vector<double>& u, const std::vector<double>& v) {
+	std::vector<double> w;
 	w.resize(u.size());
 	for (int i = 0; i < u.size(); i++) {
 		w[i] = u[i];
@@ -33,41 +33,41 @@ std::vector<float> operator*(const std::vector<float>& u, const std::vector<floa
 	}
 	return w;
 }
-std::vector<float> operator*(const std::vector<float>& u, const float& v) {
-	std::vector<float> w;
+std::vector<double> operator*(const std::vector<double>& u, const double& v) {
+	std::vector<double> w;
 	w.resize(u.size());
 	for (int i = 0; i < u.size(); i++) {
 		w[i] = v*u[i];
 	}
 	return w;
 }
-std::vector<float> operator*(const float& v, const std::vector<float>& u) {
+std::vector<double> operator*(const double& v, const std::vector<double>& u) {
 	return u*v;
 }
-float dot(const std::vector<float>& u, const std::vector<float>& v) {
-	float w = 0.0f;
+double dot(const std::vector<double>& u, const std::vector<double>& v) {
+	double w = 0.0f;
 	for (int i = 0; i < u.size() && i < v.size(); i++) {
 		w += v[i]*u[i];
 	}
 	return w;
 }
-std::vector<float> cross(const std::vector<float>& u, const std::vector<float>& v) {
-	std::vector<float> w;
+std::vector<double> cross(const std::vector<double>& u, const std::vector<double>& v) {
+	std::vector<double> w;
 	w.reserve(3);
 	for (int i = 0; i < u.size() && i < v.size(); i++) {
 		
 	}
 	return w;
 }
-float magnitude(std::vector<float>& u) {
-	float m = 0.0f;
+double magnitude(std::vector<double>& u) {
+	double m = 0.0f;
 	// std::cout << "u.size() = " << u.size() << std::endl;
 	for (int i = 0; i < u.size(); i++) {
 		m += u[i]*u[i];
 	}
 	return sqrt(m);
 }
-void normalise(std::vector<float>& u) {
-	float m = magnitude(u);
+void normalise(std::vector<double>& u) {
+	double m = magnitude(u);
 	u *= (1/m);
 }
